[UPDATE] #BTS-95 : 카프카 서비스 인터페이스 삭제
