+++ src/main/java/com/readme/payments/payments/service/PaymentsServiceImpl.java
@@ -9,7 +9,6 @@ import com.readme.payments.payments.model.PurchaseRecord;
 import com.readme.payments.payments.repository.PurchaseRepository;
 import com.readme.payments.payments.requestObject.RequestPurchase;
 import com.readme.payments.payments.responseObject.ResponsePurchase;
-import com.readme.payments.payments.service.producer.SendChargePointService;
 import com.readme.payments.payments.model.ChargeRecord;
 import com.readme.payments.payments.repository.ChargeRepository;
 import com.readme.payments.payments.requestObject.RequestApprove;
@@ -17,6 +16,7 @@ import com.readme.payments.payments.requestObject.RequestReady;
 import com.readme.payments.payments.responseObject.Message;
 import com.readme.payments.payments.responseObject.ResponseApprove;
 import com.readme.payments.payments.responseObject.ResponseReady;
+import com.readme.payments.payments.service.producer.SendChargePointService;
 import java.time.LocalDateTime;
 import java.util.Random;
 import lombok.RequiredArgsConstructor;
